requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
lightgbm>=3.3.0
ta>=0.10.0
oandapyV20>=0.6.3
websockets>=10.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
requests>=2.28.0
pyyaml>=6.0
matplotlib>=3.5.0
grpcio>=1.50.0
eth_account>=0.8.0
web3>=6.0.0
cryptography>=3.4.0
ccxt>=2.0.0
crewai>=0.1.0
fastapi>=0.95.0
uvicorn>=0.20.0
schedule>=1.2.0
psutil>=5.9.0
twilio>=8.0.0